+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- Calling `SuperTokens.signOut` now throws `STGeneralError` if the API responds with a general error status
+
 ## [12.0.0] - 2022-06-02
 
 - Adds tests for custom cookie and window handlers