+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [12.1.0] - 2022-06-15
+
 ### Changed
 
 - Calling `SuperTokens.signOut` now throws `STGeneralError` if the API responds with a general error status